+++ strahlenwerk_runtime/src/DataInterpolator.h
@@ -6,11 +6,11 @@
 #include "quat.h"
 
 namespace DataInterpolator {
-	void loadUniforms(const double time);
-	void setUniformValue(const double time, const int nthUniform, const int type, const int location);
+	void loadUniforms(const int time);
+	void setUniformValue(const int time, const int nthUniform, const int type, const int location);
 	void setValue(const int nthUniform, const int type, const int location, const int offset);
-	void setLinearValue(const int nthUniform, const int type, const int location, const int offset, const double mixT);
-	void setSplineValue(const int nthUniform, const int type, const int location, const int offset, const double mixT, const bool noFirstValue, const bool noLastValue);
+	void setLinearValue(const int nthUniform, const int type, const int location, const int offset, const float mixT);
+	void setSplineValue(const int nthUniform, const int type, const int location, const int offset, const float mixT, const bool noFirstValue, const bool noLastValue);
 };
 
 #endif // DATAINTERPOLATOR_H
